Move SendItemDurationEvents to GA4ItemTracker, call from game-end
